requirements.txt
streamlit
langchain
langchain-openai
langchain-community
python-dotenv
requests
httpx
openai
tavily-python